First contents parsed successfully XAML -> DynamicActivity
- Changed the visibility of Activity from protected to public allowed Portable.Xaml to reach the property and fill the Implementation body.

This commit allows XAML content to be sucessfully parsed to a DynamicActivity object. The <Activity><Hello World/></Activity> is parsed successfully, but then fails when invoked due to the FuncFactory.Evaluate() "XamlServices.Transform(this.Nodes.GetReader(), writer);" being called. This requires the XamlReader which fails with a "Portable.Xaml.XamlObjectWriterException: GetObject is not allowed at current state Initial" exception.

The <Activity><x:Members><...></x:Members><Assign><...></Assign></Activity> fails with a "System.Exception: should not happen" during the ActivityXamlServices.Load().